+++ src/player-util.h
@@ -80,6 +80,7 @@ void player_update_light(struct player *p);
 void player_over_exert(struct player *p, int flag, int chance, int amount);
 struct object *player_best_digger(struct player *p, bool forbid_stack);
 bool player_attack_random_monster(struct player *p);
+enum player_game_mode player_get_game_mode(const struct player *p);
 int player_check_terrain_damage(struct player *p, struct loc grid, bool actual);
 void player_take_terrain_damage(struct player *p, struct loc grid);
 struct player_shape *lookup_player_shape(const char *name);